pyproject.toml
[tool.poetry]
name = "investmentfunds"
version = "0.1.0"
description = "Investmentfunds is a suscription/cancellation investments (Challenge) from Btg RESTful API built with Python and FastAPI."
authors = ["Kevin Diaz <kevindiaz9511@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.109.1"
beanie = "^1.20.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
structlog = "^23.1.0"
uvicorn = {extras = ["standard"], version = "0.23.2"}
gunicorn = ">=21.2"
pydantic = "^2.0"
pydantic-settings = "^2.0"
loguru = "^0.7.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
asgi-lifespan = "^2.1.0"
mongomock-motor = "^0.0.34"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
ruff = "^0.0.284"
mypy = "^1.4.1"
black = "^23.7.0"
pre-commit = "^3.3.3"


[tool.isort]
profile = "black"
skip_gitignore = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black.
    "B008",  # do not perform function calls in argument defaults.
    "B010",  # Do not call `setattr` with a constant attribute value.
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.per-file-ignores]
# Ignore `F401` violations in any `__init__.py` file.
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"